requirements.txt
amqp==5.3.1
asgiref==3.8.1
billiard==4.2.1
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cryptography==44.0.1
decorator==5.2.1
defusedxml==0.7.1
Django==5.1.6
django-allauth==65.7.0
django-filter==25.1
django-ratelimit==4.1.0
django-redis==5.4.0
django_celery_results==2.6.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
djoser==2.3.1
dotenv==0.9.9
eralchemy==1.5.0
future==1.0.0
geocoder==1.38.1
geographiclib==2.0
geopy==2.4.1
greenlet==3.2.1
gunicorn==23.0.0
hiredis==3.1.0
idna==3.10
kombu==5.5.3
Markdown==3.7
oauthlib==3.2.2
packaging==24.2
pillow==11.1.0
prompt_toolkit==3.0.51
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python3-openid==3.2.0
ratelim==0.1.6
ratelimit==2.2.1
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.5.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.3.0
vine==5.1.0
wcwidth==0.2.13
